refactor(core): move AnalyzerProvider port to core.ports
Relocate the AnalyzerProvider abstract port from core.application to
core.ports — its architectural home as a driven-side interface. The return
type stays a bare Mapping[str, type] (not type[BaseAnalyzer]): core.ports may
not import core.domain where BaseAnalyzer lives, so the layer rule forbids the
concrete bind. Resolves the decision deferred in #785.

- git mv regis/core/{application,ports}/analyzer_provider.py
- importers retargeted (EntryPointAnalyzerProvider adapter + fakes/tests)
- docs synced: CLAUDE.md tree, systemPatterns layer table/diagram, decisionLog
  entry flipped from 'deferred' to 'resolved'
